chore: move parser to separate package
Move all parsing to a separate package and separate parsed statements from
executable statements. This makes the parser reusable for other purposes.

Updates #461
