+++ conn.go
@@ -28,6 +28,7 @@ import (
 	adminpb "cloud.google.com/go/spanner/admin/database/apiv1/databasepb"
 	"cloud.google.com/go/spanner/apiv1/spannerpb"
 	"github.com/googleapis/go-sql-spanner/connectionstate"
+	"github.com/googleapis/go-sql-spanner/parser"
 	"google.golang.org/api/iterator"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
@@ -245,7 +246,7 @@ type SpannerConn interface {
 var _ SpannerConn = &conn{}
 
 type conn struct {
-	parser        *statementParser
+	parser        *parser.StatementParser
 	connector     *connector
 	closed        bool
 	client        *spanner.Client
@@ -259,7 +260,7 @@ type conn struct {
 
 	execSingleQuery              func(ctx context.Context, c *spanner.Client, statement spanner.Statement, bound spanner.TimestampBound, options *ExecOptions) *spanner.RowIterator
 	execSingleQueryTransactional func(ctx context.Context, c *spanner.Client, statement spanner.Statement, options *ExecOptions) (rowIterator, *spanner.CommitResponse, error)
-	execSingleDMLTransactional   func(ctx context.Context, c *spanner.Client, statement spanner.Statement, statementInfo *statementInfo, options *ExecOptions) (*result, *spanner.CommitResponse, error)
+	execSingleDMLTransactional   func(ctx context.Context, c *spanner.Client, statement spanner.Statement, statementInfo *parser.StatementInfo, options *ExecOptions) (*result, *spanner.CommitResponse, error)
 	execSingleDMLPartitioned     func(ctx context.Context, c *spanner.Client, statement spanner.Statement, options *ExecOptions) (int64, error)
 
 	// state contains the current ConnectionState for this connection.
@@ -314,7 +315,7 @@ func (c *conn) setCommitResponse(commitResponse *spanner.CommitResponse) {
 	_ = propertyCommitTimestamp.SetValue(c.state, &commitResponse.CommitTs, connectionstate.ContextUser)
 }
 
-func (c *conn) showConnectionVariable(identifier identifier) (any, bool, error) {
+func (c *conn) showConnectionVariable(identifier parser.Identifier) (any, bool, error) {
 	extension, name, err := toExtensionAndName(identifier)
 	if err != nil {
 		return nil, false, err
@@ -322,7 +323,7 @@ func (c *conn) showConnectionVariable(identifier identifier) (any, bool, error) 
 	return c.state.GetValue(extension, name)
 }
 
-func (c *conn) setConnectionVariable(identifier identifier, value string, local bool) error {
+func (c *conn) setConnectionVariable(identifier parser.Identifier, value string, local bool) error {
 	extension, name, err := toExtensionAndName(identifier)
 	if err != nil {
 		return err
@@ -333,15 +334,15 @@ func (c *conn) setConnectionVariable(identifier identifier, value string, local 
 	return c.state.SetValue(extension, name, value, connectionstate.ContextUser)
 }
 
-func toExtensionAndName(identifier identifier) (string, string, error) {
+func toExtensionAndName(identifier parser.Identifier) (string, string, error) {
 	var extension string
 	var name string
-	if len(identifier.parts) == 1 {
+	if len(identifier.Parts) == 1 {
 		extension = ""
-		name = identifier.parts[0]
-	} else if len(identifier.parts) == 2 {
-		extension = identifier.parts[0]
-		name = identifier.parts[1]
+		name = identifier.Parts[0]
+	} else if len(identifier.Parts) == 2 {
+		extension = identifier.Parts[0]
+		name = identifier.Parts[1]
 	} else {
 		return "", "", status.Errorf(codes.InvalidArgument, "invalid variable name: %s", identifier)
 	}
@@ -796,7 +797,7 @@ func (c *conn) Prepare(query string) (driver.Stmt, error) {
 
 func (c *conn) PrepareContext(_ context.Context, query string) (driver.Stmt, error) {
 	execOptions := c.options( /* reset = */ true)
-	parsedSQL, args, err := c.parser.parseParameters(query)
+	parsedSQL, args, err := c.parser.ParseParameters(query)
 	if err != nil {
 		return nil, err
 	}
@@ -805,13 +806,17 @@ func (c *conn) PrepareContext(_ context.Context, query string) (driver.Stmt, err
 
 func (c *conn) QueryContext(ctx context.Context, query string, args []driver.NamedValue) (driver.Rows, error) {
 	// Execute client side statement if it is one.
-	clientStmt, err := c.parser.parseClientSideStatement(c, query)
+	clientStmt, err := c.parser.ParseClientSideStatement(query)
 	if err != nil {
 		return nil, err
 	}
 	execOptions := c.options( /* reset = */ clientStmt == nil)
 	if clientStmt != nil {
-		return clientStmt.QueryContext(ctx, execOptions, args)
+		execStmt, err := createExecutableStatement(clientStmt)
+		if err != nil {
+			return nil, err
+		}
+		return execStmt.queryContext(ctx, c, execOptions)
 	}
 
 	return c.queryContext(ctx, query, execOptions, args)
@@ -830,14 +835,14 @@ func (c *conn) queryContext(ctx context.Context, query string, execOptions *Exec
 	if err != nil {
 		return nil, err
 	}
-	statementType := c.parser.detectStatementType(query)
+	statementType := c.parser.DetectStatementType(query)
 	// DDL statements are not supported in QueryContext so fail early.
-	if statementType.statementType == statementTypeDdl {
+	if statementType.StatementType == parser.StatementTypeDdl {
 		return nil, spanner.ToSpannerError(status.Errorf(codes.FailedPrecondition, "QueryContext does not support DDL statements, use ExecContext instead"))
 	}
 	var iter rowIterator
 	if c.tx == nil {
-		if statementType.statementType == statementTypeDml {
+		if statementType.StatementType == parser.StatementTypeDml {
 			// Use a read/write transaction to execute the statement.
 			var commitResponse *spanner.CommitResponse
 			iter, commitResponse, err = c.execSingleQueryTransactional(ctx, c.client, stmt, execOptions)
@@ -878,13 +883,17 @@ func (c *conn) queryContext(ctx context.Context, query string, execOptions *Exec
 
 func (c *conn) ExecContext(ctx context.Context, query string, args []driver.NamedValue) (driver.Result, error) {
 	// Execute client side statement if it is one.
-	stmt, err := c.parser.parseClientSideStatement(c, query)
+	stmt, err := c.parser.ParseClientSideStatement(query)
 	if err != nil {
 		return nil, err
 	}
 	execOptions := c.options( /*reset = */ stmt == nil)
 	if stmt != nil {
-		return stmt.ExecContext(ctx, execOptions, args)
+		execStmt, err := createExecutableStatement(stmt)
+		if err != nil {
+			return nil, err
+		}
+		return execStmt.execContext(ctx, c, execOptions)
 	}
 	return c.execContext(ctx, query, execOptions, args)
 }
@@ -893,9 +902,9 @@ func (c *conn) execContext(ctx context.Context, query string, execOptions *ExecO
 	// Clear the commit timestamp of this connection before we execute the statement.
 	c.clearCommitResponse()
 
-	statementInfo := c.parser.detectStatementType(query)
+	statementInfo := c.parser.DetectStatementType(query)
 	// Use admin API if DDL statement is provided.
-	if statementInfo.statementType == statementTypeDdl {
+	if statementInfo.StatementType == parser.StatementTypeDdl {
 		// Spanner does not support DDL in transactions, and although it is technically possible to execute DDL
 		// statements while a transaction is active, we return an error to avoid any confusion whether the DDL
 		// statement is executed as part of the active transaction or not.
@@ -1297,7 +1306,7 @@ func queryInNewRWTransaction(ctx context.Context, c *spanner.Client, statement s
 
 var errInvalidDmlForExecContext = spanner.ToSpannerError(status.Error(codes.FailedPrecondition, "Exec and ExecContext can only be used with INSERT statements with a THEN RETURN clause that return exactly one row with one column of type INT64. Use Query or QueryContext for DML statements other than INSERT and/or with THEN RETURN clauses that return other/more data."))
 
-func execInNewRWTransaction(ctx context.Context, c *spanner.Client, statement spanner.Statement, statementInfo *statementInfo, options *ExecOptions) (*result, *spanner.CommitResponse, error) {
+func execInNewRWTransaction(ctx context.Context, c *spanner.Client, statement spanner.Statement, statementInfo *parser.StatementInfo, options *ExecOptions) (*result, *spanner.CommitResponse, error) {
 	var res *result
 	options.QueryOptions.LastStatement = true
 	fn := func(ctx context.Context, tx *spanner.ReadWriteTransaction) error {
@@ -1315,7 +1324,7 @@ func execInNewRWTransaction(ctx context.Context, c *spanner.Client, statement sp
 	return res, &resp, nil
 }
 
-func execTransactionalDML(ctx context.Context, tx spannerTransaction, statement spanner.Statement, statementInfo *statementInfo, options spanner.QueryOptions) (*result, error) {
+func execTransactionalDML(ctx context.Context, tx spannerTransaction, statement spanner.Statement, statementInfo *parser.StatementInfo, options spanner.QueryOptions) (*result, error) {
 	var rowsAffected int64
 	var lastInsertId int64
 	var hasLastInsertId bool
@@ -1327,7 +1336,7 @@ func execTransactionalDML(ctx context.Context, tx spannerTransaction, statement 
 	}
 	if len(it.Metadata.RowType.Fields) != 0 && !(len(it.Metadata.RowType.Fields) == 1 &&
 		it.Metadata.RowType.Fields[0].Type.Code == spannerpb.TypeCode_INT64 &&
-		statementInfo.dmlType == dmlTypeInsert) {
+		statementInfo.DmlType == parser.DmlTypeInsert) {
 		return nil, errInvalidDmlForExecContext
 	}
 	if err != iterator.Done {
