+++ driver.go
@@ -40,6 +40,7 @@ import (
 	"github.com/google/uuid"
 	"github.com/googleapis/gax-go/v2"
 	"github.com/googleapis/go-sql-spanner/connectionstate"
+	"github.com/googleapis/go-sql-spanner/parser"
 	"google.golang.org/api/iterator"
 	"google.golang.org/api/option"
 	"google.golang.org/api/option/internaloption"
@@ -529,7 +530,7 @@ type connector struct {
 	adminClient    *adminapi.DatabaseAdminClient
 	adminClientErr error
 	connCount      int32
-	parser         *statementParser
+	parser         *parser.StatementParser
 }
 
 func newOrCachedConnector(d *Driver, dsn string) (*connector, error) {
@@ -732,7 +733,7 @@ func openDriverConn(ctx context.Context, c *connector) (driver.Conn, error) {
 	logger := c.logger.With("connId", connId)
 	connectionStateType := c.connectorConfig.ConnectionStateType
 	if connectionStateType == connectionstate.TypeDefault {
-		if c.parser.dialect == databasepb.DatabaseDialect_POSTGRESQL {
+		if c.parser.Dialect == databasepb.DatabaseDialect_POSTGRESQL {
 			connectionStateType = connectionstate.TypeTransactional
 		} else {
 			connectionStateType = connectionstate.TypeNonTransactional
@@ -806,7 +807,7 @@ func (c *connector) increaseConnCount(ctx context.Context, databaseName string, 
 			} else if c.connectorConfig.StatementCacheSize == 0 {
 				cacheSize = defaultStatementCacheSize
 			}
-			c.parser, err = newStatementParser(dialect, cacheSize)
+			c.parser, err = parser.NewStatementParser(dialect, cacheSize)
 			if err != nil {
 				closeClient()
 				return err
