+++ parser/README.md
@@ -1,0 +1,3 @@
+# Spanner Statement Parser
+
+This package is an internal package and can make breaking changes without prior notice.
