+++ parser/statement_cache.go
@@ -1,0 +1,8 @@
+package parser
+
+type statementsCacheEntry struct {
+	sql             string
+	params          []string
+	info            *StatementInfo
+	parsedStatement ParsedStatement
+}
